requirements.txt
django==4.2.17
djangorestframework
mysqlclient
python-dotenv
dj-database-url
